CXO-2951: Refactor device messages: make sure all constructors use long for id field

+++ upl/com.energyict.mdc.upl.impl/protocolcommon/src/main/java/com/energyict/protocolimplv2/messages/ActivityCalendarDeviceMessage.java
@@ -175,7 +175,7 @@ public enum ActivityCalendarDeviceMessage implements DeviceMessageSpecSupplier {
     private final long id;
     private final String defaultNameTranslation;
 
-    ActivityCalendarDeviceMessage(int id, String defaultNameTranslation) {
+    ActivityCalendarDeviceMessage(long id, String defaultNameTranslation) {
         this.id = id;
         this.defaultNameTranslation = defaultNameTranslation;
     }
