+++ upl/com.energyict.mdc.upl.impl/protocolcommon/src/main/java/com/energyict/protocolimplv2/messages/LoadProfileMessage.java
@@ -127,7 +127,7 @@ public enum LoadProfileMessage implements DeviceMessageSpecSupplier {
     private final long id;
     private final String defaultNameTranslation;
 
-    LoadProfileMessage(int id, String defaultNameTranslation) {
+    LoadProfileMessage(long id, String defaultNameTranslation) {
         this.id = id;
         this.defaultNameTranslation = defaultNameTranslation;
     }
