+++ upl/com.energyict.mdc.upl.impl/protocolcommon/src/main/java/com/energyict/protocolimplv2/messages/TotalizersConfigurationDeviceMessage.java
@@ -47,7 +47,7 @@ public enum TotalizersConfigurationDeviceMessage implements DeviceMessageSpecSup
     private final long id;
     private final String defaultNameTranslation;
 
-    TotalizersConfigurationDeviceMessage(int id, String defaultNameTranslation) {
+    TotalizersConfigurationDeviceMessage(long id, String defaultNameTranslation) {
         this.id = id;
         this.defaultNameTranslation = defaultNameTranslation;
     }
